Update solver.py
Only trying to click "recaptcha-audio-button", after which directly go to "_solve_audio_challenge". 

Reasons:
The current way of trying to click "recaptcha-image-button" first can easily trigger the"Google automated queries detetion" if the audio challenge is shown in the first place instead an image challenge.

The proposed changes can simplfied the process, as either successfully or unsuccessfully clicking "recaptcha-audio-button", the audio challenge is ready to be solved.